Remove Retry sending if sending fails in conn_pool
Summary:
Part of the changes we want to make for better backoff. There are
backoff-ish things in differents parts of
the code(scribe_server, store_queue, store and conn_pool) which makes it
hard to reason about backoff scenarios. Reasoning would be much easier
if the response is centralized
to only a few places in the code. This code removed in this diff is
fairly useless. We should have our backoff handling code deal react to
sending failures rather than retrying a second time

Reviewed By: agiardullo

Test Plan:
Test with the stress.php test setup for simulating backoff that
this code works

Revert: OK