Exponential backoff Algorithm for scribe
Summary:
This is the exponential backoff algorithm for scribe. Basically
we implement a simple additive increase(reduce retryInterval) and
multiplicatve decrease(increase retryInterval) scheme. This algorithm
has been chosen because it is known to provide good congestion control
while ensuring fairness when large number of senders send to a
particular receiver.

This adds a new config parameter "adaptive_backoff" which must be set
to yes for using this algorithm. This also adds 3 other parameters:
 min_retry_interval, max_retry_interval and random_offset_range. The
 third parameter is used to add a random number to the retry_interval
 value the backoff generates to prevent the thundering herd problem.

 I have also provided default values for these parameters which may need
 some tuning

 Reviewed By: agiardullo

 Test Plan:
 Tested with the scenario in which we use to simulate backoff. The
 configs and readme provided with this test describe the test.  Involves
 two dev
 servers as senders and one devserver as receiver. superstress.php was
 run on the senders. The receiver was started a few minutes later which
 resulted in denied for queue sizes. More test results provided below.