Handle failure of listImpl in HDfsFile
Summary:
we found that when there are problems w/hdfs and an ls fails,
scribe treats this as ls returning no files and creates the default file
with index
(00000). Now this method throws an exception and the exception is
suitably handled.

Test Plan:
Testsuite. Simulate a backup scenario and see that it works

DiffCamp Revision: 113049
Reviewed By: zshao
CC: agiardullo, zshao, groys, scribe-dev@lists
Tasks:
#206325: scribe : handle list errors from hdfs

Revert Plan:
OK

git-svn-id: svn+ssh://tubbs/svnapps/fbomb/branches/scribe-os/fbcode/scribe@27540 2248de34-8caa-4a3c-bc55-5e52d9d7b73a