Remove stores: category, multifile, multithriftfile
Summary:
these are very old and are designed to provide a new thread per
category. I dont know whether they interact well with the paramter
new_thread_per_category. These stores are undocumented so nobody should
be using them anyway. This code has not been updated in 2 years, So the
chances of its working with our recent changes are slim. We will get rid
of this and make new_thread_per_category a store level parameter if need
be

Test Plan:
Testsuite