Fix improper handling of corrupted data when reading and writing buffer files.
Summary:
On read part:
  originally, scribe will discard any data in the file from the point where
corruption is detected, but with this handling scheme, part of the corrupted
data can still be read and sent to primary. An example is illustrated as
follows:

  <--- GOOD FRAME --><--- PARTIAL FRAME --| <--- NEXT GOOD FRAME --->
                     ^                    ^ the corrupted frame got truncated
here
                     the frame size header of the corrupted frame is OK

in the above example, when the corrupted frame is read, StdFile::readNext() doesn't
know it is corrupted (since the frame size header is OK), and reads the partial frame
together with the leading part of the next good frame. When we feed this corrupted
frame to TBinaryProtocol to deserialize it, thrift throws an exception. This exception
was not properly handled, so that scribe failed to pass this buffer file, resulting
in buffer file piling up (is buffer streaming disabled).

This diff catches the exception and discards all data from the beginning of the
corrupted frame as the old scribe does. Ideally, we would like a smarter handling
policy like local scanning to adjust the read pointer to the beginning of the next
good frame, but that would require significant change to FileInterface. As discussed
offline, we would like to create a new file type (e.g. BufferFile) later to do that.

On write part:
  the original scribe doesn't rotate file on write failure. Next time handleMessages()
is called, the corrupted file is re-opened, and good data are appended to the end. This
result in loss of all data after the corruption point. We change the policy to rotate
file on write failures, so that only the last few bytes of the corrupted file will be
lost.

This diff also fix the badalloc exception described in the task. That exception
happens because we keep using TMemoryBuffer for deserialization without
resetting it. This causes its r/w pointers to keep moving forward without
releasing any buffer space before rBase_, which effectively is a memory leak.

Test Plan:
For the read part:
  scribe scribe with a midtier-writer configuration;
  do stress test;
  scribe_ctrl stop the writer scribe; the mid-tier scribe should begin
generating buffers;
  open the oldest buffer file and change some bytes in it so that it is
corrupted*.
  start the writer scribe again so that buffers are replayed.
  scribe should complain errors and skip rest of the corrupted file.

For the write part:
  mkdir /tmp/tmpfs
  sudo mount -t tmpfs -o size=20m tmpfs /tmp/tmpfs
  start a mid-tier scribe;
  run stresstest.php;
  when the file size quota is reached, scribe should rotate to next file
instead of
trying to re-open the original file. There may be some partial data at the end
of the original
file.
  sudo mount -t tmpfs -o remount -o size=100m tmpfs /tmp/tmpfs
  scribe should flush the remaining data